+++ Manager/OroFieldSelectManager.php
@@ -4,6 +4,7 @@ namespace Synolia\Bundle\OroneoBundle\Manager;
 
 use Oro\Bundle\EntityConfigBundle\Config\ConfigManager;
 use Oro\Bundle\LocaleBundle\Helper\LocalizationHelper;
+use Oro\Bundle\CatalogBundle\Entity\Repository\CategoryRepository;
 
 /**
  * Class OroFieldSelectManager
@@ -17,16 +18,24 @@ class OroFieldSelectManager
     /** @var LocalizationHelper $localizationHelper */
     protected $localizationHelper;
 
+    /** @var CategoryRepository $categoryRepository */
+    protected $categoryRepository;
+
     /**
      * OroFieldSelectManager constructor.
      *
      * @param ConfigManager      $configManager
      * @param LocalizationHelper $localizationHelper
+     * @param CategoryRepository $categoryRepository
      */
-    public function __construct(ConfigManager $configManager, LocalizationHelper $localizationHelper)
-    {
+    public function __construct(
+        ConfigManager $configManager,
+        LocalizationHelper $localizationHelper,
+        CategoryRepository $categoryRepository
+    ) {
         $this->configManager      = $configManager;
         $this->localizationHelper = $localizationHelper;
+        $this->categoryRepository = $categoryRepository;
     }
 
     /**
@@ -36,9 +45,14 @@ class OroFieldSelectManager
      */
     public function getChoices($className)
     {
-        $choices = [];
-        $product  = $this->configManager->getConfigEntityModel($className);
-        foreach ($product->getFields() as $field) {
+        $choices     = [];
+        $configModel = $this->configManager->getConfigEntityModel($className);
+
+        if (!$configModel) {
+            return [];
+        }
+
+        foreach ($configModel->getFields() as $field) {
             $choices[$field->getFieldName()] = $field->getFieldName();
         }
 
@@ -56,7 +70,7 @@ class OroFieldSelectManager
      */
     public function getLocalizationChoices()
     {
-        $localizations = $this->localizationHelper->getAll();
+        $localizations = $this->localizationHelper->getLocalizations();
         $choices       = [
             'default' => 'synolia.oroneo.configuration_page.mapping.localization.default.label',
         ];
@@ -66,5 +80,27 @@ class OroFieldSelectManager
         }
 
         return $choices;
+    }
+
+    /**
+     * Return array of Categories.
+     *
+     * @return array
+     */
+    public function getCategoriesChoices()
+    {
+        $rootCategory = $this->categoryRepository->getMasterCatalogRoot();
+
+        $children = $this->categoryRepository->getAllChildCategories($rootCategory);
+
+        $select[$rootCategory->getId()] = $rootCategory->getDefaultTitle()->getString();
+
+        foreach ($children as $child) {
+            $space = str_repeat('--', $child->getLevel());
+            $select[$child->getId()] = $space.' '.$child->getDefaultTitle()->getString();
+
+        }
+
+        return $select;
     }
 }